requirements.txt
numpy==2.4.2
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
opencv_python==4.13.0.90
opencv_python_headless==4.12.0.88
pyorbbecsdk2==2.0.15
tomli==2.0.1
